Added logic to create only necessary queues
Previously, the code would treat the presentation queue and graphics queue as two indepedent entities regardless if the queue family supports both functionalities. By including a few checks, the program can now decide if the queue families share functionality and, if so, to create only the necessary amount of queues.